worldstate: move merklenode and merklevalue to submod node
